From Now on my Stack just differentiate between doubles and size_t's internally! also added unique_ptr array of bytes into the VM but it is not used yet

+++ VM/ByteCode.h
@@ -12,7 +12,7 @@ namespace vm
 
 	class VM;
 
-	enum ByteCodes : long {
+	enum ByteCodes : unsigned long long {
 		IADD = 1,
 		ISUB,
 		IMUL,
